Use general g:render instead of f:table, f:all or f:display
because the f:* is buggy and messy in the usage.
g:render is the most general way to use a template in Grails.
And you can specify a template name explicitly as you want. It's very handy.
f:widget and f:displayWidget are useful even in the case.
I think it's one of the best way for using fields plugin.